% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as_vector}
\alias{as_vector}
\title{Cast to list vector generic}
\usage{
as_vector(x, pattern, which, ...)
}
\arguments{
\item{x}{an S3-object to convert.}

\item{pattern}{an optional \link{regular expression}.  Only names
matching \code{pattern} are returned. \code{\link{glob2rx}} can be
used to convert wildcard patterns to regular expressions.}

\item{which}{a character vector, either \code{"global"} to print all unrecognized global variables,
\code{"import"} to print all detected imported functions and variables, or both (default).}

\item{...}{additional arguments to configure the returned output.}
}
\value{
a list of character vectors.
}
\description{
\code{as_vector} is a generic function to cast objects returned by \code{\link{checkglobals}},
\code{\link{check_pkg}} or \code{\link{check_source}} to list vectors. The function invokes particular
\emph{methods} which depend on the \code{\link{class}} of the first argument.
}
